requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
Django==5.2.6
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
jiter==0.11.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
openai==1.107.2
packaging==25.0
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.3
timm==1.0.12
tokenizers==0.19.1
torch==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.67.1
transformers==4.44.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
